% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aop_merge}
\alias{aop_merge}
\title{Merge land use and access data}
\usage{
aop_merge(aop_landuse, aop_access)
}
\arguments{
\item{aop_landuse}{A \code{data.table} of aop land use data}

\item{aop_access}{A \code{data.table} of aop access data}
}
\value{
Returns a \code{data.table} with landuse and access data
}
\description{
Merges landuse and access data
}
\keyword{internal}
