% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aop_spatial_join}
\alias{aop_spatial_join}
\title{Spatial join of AOP data}
\usage{
aop_spatial_join(aop_df, aop_sf)
}
\arguments{
\item{aop_df}{A \code{data.frame} of aop data}

\item{aop_sf}{A spatial \code{sf} of aop data}
}
\description{
Spatial join of AOP data
}
\seealso{
Other support functions: 
\code{\link{download_gpkg}()},
\code{\link{load_gpkg}()},
\code{\link{rm_accent}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
