% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aopdata.R
\docType{package}
\name{aopdata}
\alias{aopdata}
\alias{aopdata-package}
\title{aopdata: Data from the 'Access to Opportunities Project (AOP)'}
\description{
Download data from the 'Access to Opportunities Project (AOP)'. The aopdata
package brings annual estimates of access to employment, health, education
and social assistance services by transport mode,  as well as data  on the
spatial distribution of population, jobs,  health care, schools and social
assistance facilities at a fine spatial resolution for all cities included in
the project. More  info on the AOP website \url{https://www.ipea.gov.br/acessooportunidades/en/}.
}
\section{Usage}{

Please check the vignettes on the \href{https://ipeagit.github.io/aopdata/}{website}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://ipeagit.github.io/aopdata/}
  \item \url{https://github.com/ipeaGIT/aopdata}
  \item Report bugs at \url{https://github.com/ipeaGIT/aopdata/issues}
}

}
\author{
\strong{Maintainer}: Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})

Authors:
\itemize{
  \item Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})
  \item Marcus Saraiva \email{marcus.saraiva@gmail.com} (\href{https://orcid.org/0000-0001-6218-2338}{ORCID})
  \item Carlos Kaue Vieira Braga \email{kaue@kauebraga.dev} (\href{https://orcid.org/0000-0002-6104-7297}{ORCID})
}

Other contributors:
\itemize{
  \item Diego Bogado Tomasiello \email{diegobt86@gmail.com} [contributor]
  \item Joao Bazzo \email{joao.bazzo@gmail.com} [contributor]
  \item Ipea - Institute for Applied Economic Research [copyright holder, funder]
}

}
\keyword{internal}
