% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aopdata_dictionary.R
\name{aopdata_dictionary}
\alias{aopdata_dictionary}
\title{aopdata data dictionary}
\usage{
aopdata_dictionary(lang = "en")
}
\arguments{
\item{lang}{Character. Language of data dictionary. It can be either \code{"en"}
for English (default) or \code{"pt"} for Portuguese.}
}
\value{
Opens aopdata data dictionary on a web browser
}
\description{
Opens aopdata data dictionary on a web browser. This function requires
internet connection.
}
\examples{
\dontrun{ if (interactive()) {

# Data dictionary in English
aopdata_dictionary(lang='en')

#' # Data dictionary in Portuguese
aopdata_dictionary(lang='pt')

}}
}
\seealso{
Other support functions: 
\code{\link{aop_merge}()},
\code{\link{aop_spatial_join}()},
\code{\link{check_connection}()},
\code{\link{download_data}()},
\code{\link{load_data}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
