% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_downloaded_obj}
\alias{check_downloaded_obj}
\title{Check if object has been downloaded/created in global environment}
\usage{
check_downloaded_obj(obj)
}
\arguments{
\item{obj}{Any type of object}
}
\value{
Returns \code{NULL} with an informative message
}
\description{
Check if object has been downloaded/created in global environment
}
\keyword{internal}
