% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_data}
\alias{download_data}
\title{Download data to a temporary directory.}
\usage{
download_data(file_url, progress_bar = showProgress)
}
\arguments{
\item{file_url}{A string with the file_url address of aop dataset}

\item{progress_bar}{Logical. Defaults to (TRUE) display progress bar}
}
\value{
No visible output. The downloaded file (either an \code{sf} or a
\code{data.frame}) is saved to a temporary directory.
}
\description{
Save requested data (either an \code{sf} or a \code{data.frame})
to a temporary directory.
}
\seealso{
Other support functions: 
\code{\link{aop_merge}()},
\code{\link{aop_spatial_join}()},
\code{\link{aopdata_dictionary}()},
\code{\link{check_connection}()},
\code{\link{load_data}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
