% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metadata.R
\name{download_metadata}
\alias{download_metadata}
\title{Support function to download metadata internally used in aop}
\usage{
download_metadata()
}
\description{
Support function to download metadata internally used in aop
}
\examples{
\donttest{
df <- download_metadata()
}
}
\concept{general support functions}
