% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_access.R
\name{read_access}
\alias{read_access}
\title{Download accessibility estimates, population and land use data}
\usage{
read_access(
  city,
  mode = "walk",
  year = 2019,
  geometry = FALSE,
  showProgress = TRUE
)
}
\arguments{
\item{city}{Character. A city name or three-letter abbreviation. If
\code{city="all"}, results for all cities are loaded.}

\item{mode}{Character. A transport mode. Modes available include
'public_transport', 'bicycle', or 'walk' (the default).}

\item{year}{Numeric. A year number in YYYY format. Default set to 2019, the
only year currently available.}

\item{geometry}{Logical. if FALSE (the default), returns a regular data.table
of aop data. if TRUE, returns a an \verb{sf data.frame} with simple
feature geometry of spatial hexagonal grid H3. See details in
\link{read_grid}.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\value{
A \code{data.frame} object
}
\description{
Download annual estimates of access to employment, health and education
services by transport mode, as well as data on the spatial distribution of
population, schools and healthcare facilities at a fine spatial resolution
for all cities included in the study.
}
\examples{
\donttest{
# Read accessibility estimates of a single city
bho <- read_access(city = 'Belo Horizonte', mode = 'walk', year = 2019)
bho <- read_access(city = 'bho', mode = 'walk', year = 2019)

# Read accessibility estimates for all cities
all <- read_access(city = 'all', mode = 'public_transport', year = 2019)
}
}
\concept{accessibility data functions}
