% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_landuse.R
\name{read_landuse}
\alias{read_landuse}
\title{Download land use and population data}
\usage{
read_landuse(city = NULL, year = 2019, geometry = FALSE, showProgress = TRUE)
}
\arguments{
\item{city}{Character. A city name or three-letter abbreviation. If
\code{city="all"}, results for all cities are loaded.}

\item{year}{Numeric. A year number in YYYY format. Default set to 2019, the
only year currently available.}

\item{geometry}{Logical. If \code{FALSE} (the default), returns a regular data.table
of aop data. If \code{TRUE}, returns a \verb{sf data.frame} with simple
feature geometry of spatial hexagonal grid H3. See details in
\link{read_grid}.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\value{
A \code{data.frame} object or an \verb{sf data.frame} object
}
\description{
Download data on the spatial distribution of population, schools and
healthcare facilities at a fine spatial resolution for cities included in the
study. The data comes aggregated on a hexagonal grid based on the global H3
index at resolution 9, with a size of 357 meters (short diagonal) and an area
of 0.74 km2. More information about H3 at \url{https://h3geo.org/docs/core-library/restable/}.

See documentation 'Details' for the data dictionary.
}
\section{Data dictionary:}{
\tabular{llll}{
   \strong{Data type} \tab \strong{column} \tab \strong{Description} \tab \strong{Value} \cr
   geographic \tab \code{abbrev_muni} \tab Abbreviation of city name (3 letters) \tab  \cr
   geographic \tab \code{name_muni} \tab City name \tab  \cr
   geographic \tab \code{code_muni} \tab 7-digit code of each city \tab  \cr
   geographic \tab \code{id_hex} \tab Unique id of hexagonal cell \tab  \cr
   sociodemographic \tab \code{P001} \tab Total number of residents \tab  \cr
   sociodemographic \tab \code{P002} \tab Number of white residents \tab  \cr
   sociodemographic \tab \code{P003} \tab Number of black residents \tab  \cr
   sociodemographic \tab \code{P004} \tab Number of indigenous residents \tab  \cr
   sociodemographic \tab \code{P005} \tab Number of asian-descendent residents \tab  \cr
   sociodemographic \tab \code{P010} \tab Number of residents with age between 0 and 5 \tab  \cr
   sociodemographic \tab \code{P011} \tab Number of residents with age between 6 and 14 \tab  \cr
   sociodemographic \tab \code{P012} \tab Number of residents with age between 15 and 18 \tab  \cr
   sociodemographic \tab \code{P013} \tab Number of residents with age between 19 and 24 \tab  \cr
   sociodemographic \tab \code{P014} \tab Number of residents with age between 25 and 39 \tab  \cr
   sociodemographic \tab \code{P015} \tab Number of residents with age between 40 and 69 \tab  \cr
   sociodemographic \tab \code{P016} \tab Number of residents with age 70 or more \tab  \cr
   sociodemographic \tab \code{R001} \tab Average household income per capita \tab R$ (Brazilian Reais), values in 2010 \cr
   sociodemographic \tab \code{R002} \tab Income quintile group \tab 1 (poorest), 2, 3, 4, 5 (richest) \cr
   sociodemographic \tab \code{R003} \tab Income decile group \tab 1 (poorest), 2, 3, 4, 5, 6, 7, 8, 9, 10 (richest) \cr
   land use \tab \code{T001} \tab Total number of formal jobs \tab  \cr
   land use \tab \code{T001} \tab Total number of formal jobs \tab  \cr
   land use \tab \code{T002} \tab Number of formal jobs with primary education or lower \tab  \cr
   land use \tab \code{T003} \tab Number of formal jobs with secondary education \tab  \cr
   land use \tab \code{T004} \tab Number of formal jobs with tertiary education \tab  \cr
   land use \tab \code{E001} \tab Total number of public schools \tab  \cr
   land use \tab \code{E002} \tab Number of public schools - early childhood \tab  \cr
   land use \tab \code{E003} \tab Number of public schools - elementary schools \tab  \cr
   land use \tab \code{E004} \tab Number of public schools - high schools \tab  \cr
   land use \tab \code{M001} \tab Total number of kids enrolled in public schools \tab  \cr
   land use \tab \code{M002} \tab Number of kids enrolled  in public schools - early childhood \tab  \cr
   land use \tab \code{M003} \tab Number of kids enrolled  in public schools - elementary schools \tab  \cr
   land use \tab \code{M004} \tab Number of kids enrolled  in public schools - high schools \tab  \cr
   land use \tab \code{S001} \tab Total number of healthcare facilities \tab  \cr
   land use \tab \code{S002} \tab Number of healthcare facilities - low complexity \tab  \cr
   land use \tab \code{S003} \tab Number of healthcare facilities - medium complexity \tab  \cr
   land use \tab \code{S004} \tab Number of healthcare facilities - high complexity \tab  \cr
   land use \tab \code{C001} \tab Number of centers for social assistance (CRAS) \tab  \cr
}
}

\section{Cities available}{
\tabular{ll}{
   \strong{City name} \tab \strong{Three-letter abbreviation} \cr
   Belem \tab \code{bel} \cr
   Belo Horizonte \tab \code{bho} \cr
   Brasilia \tab \code{bsb} \cr
   Campinas \tab \code{cam} \cr
   Campo Grande \tab \code{cgr} \cr
   Curitiba \tab \code{cur} \cr
   Duque de Caxias \tab \code{duq} \cr
   Fortaleza \tab \code{for} \cr
   Goiania \tab \code{goi} \cr
   Guarulhos \tab \code{gua} \cr
   Maceio \tab \code{mac} \cr
   Manaus \tab \code{man} \cr
   Natal \tab \code{nat} \cr
   Porto Alegre \tab \code{poa} \cr
   Recife \tab \code{rec} \cr
   Rio de Janeiro \tab \code{rio} \cr
   Salvador \tab \code{sal} \cr
   Sao Goncalo \tab \code{sgo} \cr
   Sao Luis \tab \code{slz} \cr
   Sao Paulo \tab \code{spo} \cr
}
}

\examples{
\dontrun{ if (interactive()) {
# a single city
bho <- read_landuse(city = 'Belo Horizonte', year = 2019, showProgress = FALSE)
bho <- read_landuse(city = 'bho', year = 2019, showProgress = FALSE)

# all cities
all <- read_landuse(city = 'all', year = 2019)
}}
}
\concept{land use data functions}
