% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_metadata}
\alias{select_metadata}
\title{Select metadata}
\usage{
select_metadata(t = NULL, c = NULL, y = NULL, m = NULL)
}
\arguments{
\item{t}{Type of data: 'access' or 'grid' (passed from \code{read_} function)}

\item{c}{City (passed from \code{read_} function)}

\item{y}{Year of the dataset (passed from \code{read_} function)}

\item{m}{Transport mode (passed from \code{read_} function)}
}
\value{
A \code{data.frame} object with metadata subsetted by data type,
'city', 'year' and 'mode'
}
\description{
Subsets the metadata table by data 'type', 'city', 'year' and
'mode'
}
\examples{
\donttest{
df <- download_metadata()
}
}
\seealso{
Other support functions: 
\code{\link{aop_merge}()},
\code{\link{aop_spatial_join}()},
\code{\link{check_connection}()},
\code{\link{download_data}()},
\code{\link{load_data}()},
\code{\link{select_city_input}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
\keyword{internal}
