% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_mode_input}
\alias{select_mode_input}
\title{Select mode input}
\usage{
select_mode_input(temp_meta = temp_meta, mode = NULL)
}
\arguments{
\item{temp_meta}{A dataframe with the file_url addresses of aop datasets}

\item{mode}{Transport mode (passed by read_ function)}
}
\value{
A \code{data.frame} object with metadata subsetted by 'mode'
}
\description{
Subsets the metadata table by 'mode'.
}
\seealso{
Other support functions: 
\code{\link{aop_merge}()},
\code{\link{aop_spatial_join}()},
\code{\link{check_connection}()},
\code{\link{download_data}()},
\code{\link{load_data}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
