% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aop_merge}
\alias{aop_merge}
\title{Merge land use and access data}
\usage{
aop_merge(aop_landuse, aop_access)
}
\arguments{
\item{aop_landuse}{A \code{data.table} of aop land use data}

\item{aop_access}{A \code{data.table} of aop access data}
}
\value{
Returns a \code{data.table} with landuse and access data
}
\description{
Merges landuse and access data
}
\seealso{
Other support functions: 
\code{\link{aop_spatial_join}()},
\code{\link{download_data}()},
\code{\link{load_data}()},
\code{\link{rm_accent}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
