% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aop_spatial_join}
\alias{aop_spatial_join}
\title{Spatial join of AOP data}
\usage{
aop_spatial_join(aop_df, aop_sf)
}
\arguments{
\item{aop_df}{A \code{data.frame} of aop data}

\item{aop_sf}{A spatial \code{sf} of aop data}
}
\value{
Returns a \verb{data.frame sf} with access/land use data and grid geometries
}
\description{
Merges land use or access data with H3 grid geometries
}
\keyword{internal}
