% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aop_spatial_join}
\alias{aop_spatial_join}
\title{Spatial join of AOP data}
\usage{
aop_spatial_join(aop_df, aop_sf)
}
\arguments{
\item{aop_df}{A \code{data.frame} of aop data}

\item{aop_sf}{A spatial \code{sf} of aop data}
}
\value{
Returns a \verb{data.frame sf} with access/landuse data and grid geometries
}
\description{
Merges landuse or access data with H3 grid geometries
}
\seealso{
Other support functions: 
\code{\link{aop_merge}()},
\code{\link{check_connection}()},
\code{\link{download_data}()},
\code{\link{load_data}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
