% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aopdata_dictionary.R
\name{aopdata_dictionary}
\alias{aopdata_dictionary}
\title{aopdata data dictionary}
\usage{
aopdata_dictionary(lang = "en")
}
\arguments{
\item{lang}{Character. Language of data dictionary. It can be either \code{"en"}
for English (default) or \code{"pt"} for Portuguese.}
}
\value{
Opens aopdata data dictionary on a web browser
}
\description{
Opens aopdata data dictionary on a web browser. This function requires
internet connection.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Data dictionary in English
aopdata_dictionary(lang='en')

# Data dictionary in Portuguese
aopdata_dictionary(lang='pt')
\dontshow{\}) # examplesIf}
}
\concept{Data dictionary}
