% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_connection}
\alias{check_connection}
\title{Check internet connection with Ipea server}
\usage{
check_connection(
  file_url = "https://www.ipea.gov.br/geobr/aopdata/metadata/metadata.csv"
)
}
\arguments{
\item{file_url}{A string with the file_url address of an aop dataset}
}
\value{
Logical. \code{TRUE} if url is working, \code{FALSE} if not.
}
\description{
Checks if there is an internet connection with Ipea server to download aop data.
}
\seealso{
Other support functions: 
\code{\link{aop_merge}()},
\code{\link{aop_spatial_join}()},
\code{\link{download_data}()},
\code{\link{load_data}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
\keyword{internal}
