% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metadata.R
\name{download_metadata}
\alias{download_metadata}
\title{Support function to download metadata internally used in 'aopdata'}
\usage{
download_metadata()
}
\value{
A \code{data.frame} object with metadata and url of data sets
}
\description{
Support function to download metadata internally used in 'aopdata'
}
\examples{
df <- download_metadata()

}
\concept{general support functions}
