% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load data from tempdir to global environment}
\usage{
load_data(temps = NULL)
}
\arguments{
\item{temps}{The address of a data file stored in tempdir. Defaults to NULL}
}
\value{
Returns either an \code{sf} or a \code{data.frame}, depending of the data set
that was downloaded
}
\description{
Reads data from tempdir to global environment.
}
\keyword{internal}
