% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load data from tempdir to global environment}
\usage{
load_data(file_url, temps = NULL)
}
\arguments{
\item{file_url}{A string with the file_url address of a aop dataset}

\item{temps}{The address of a data file stored in tempdir. Defaults to NULL}
}
\value{
Returns either an \code{sf} or a \code{data.frame}, dependeding of the data set
that was downloaded
}
\description{
Reads data from tempdir to global environment.
}
\seealso{
Other support functions: 
\code{\link{aop_merge}()},
\code{\link{aop_spatial_join}()},
\code{\link{check_connection}()},
\code{\link{download_data}()},
\code{\link{rm_accent}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
