% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_access.R
\name{read_access}
\alias{read_access}
\title{Download accessibility estimates}
\usage{
read_access(city, mode, year, geometry = FALSE, showProgress = TRUE)
}
\arguments{
\item{city}{A city name or three-letter abbreviation}

\item{mode}{A transport mode.}

\item{year}{A date number in YYYY format.}

\item{geometry}{if FALSE (the default), returns a regular data.table of aop
data. if TRUE, returns a an \verb{sf data.frame} with simple
feature geometry of spatial hexagonal grid H3.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\value{
A \code{data.frame} object
}
\description{
Download accessibility estimates
}
\examples{
\donttest{
# Read accessibility estimates of a single city
bho <- read_access(city = 'Belo Horizonte', mode = 'walk', year = 2019)
bho <- read_access(city = 'bho', mode = 'walk', year = 2019)

}
}
\concept{accessibility data functions}
