% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_access.R
\name{read_access}
\alias{read_access}
\title{Download accessibility estimates, population and land use data}
\usage{
read_access(
  city,
  mode = "walk",
  peak = TRUE,
  year = 2019,
  geometry = FALSE,
  showProgress = TRUE
)
}
\arguments{
\item{city}{Character. A city name or three-letter abbreviation. If
\code{city="all"}, results for all cities are loaded.}

\item{mode}{Character. A transport mode. Modes available include
'public_transport', 'bicycle', or 'walk' (the default).}

\item{peak}{Logical. If \code{TRUE} (the default), returns accessibility estimates
during peak time, between 6am and 8am. If \code{FALSE}, returns
accessibility during off-peak, between 2pm and 4am. This argument
only takes effect when \code{mode = public_transport}.}

\item{year}{Numeric. A year number in YYYY format. Default set to 2019, the
only year currently available.}

\item{geometry}{Logical. If \code{FALSE} (the default), returns a regular data.table
of aop data. If \code{TRUE}, returns a an \verb{sf data.frame} with simple
feature geometry of spatial hexagonal grid H3. See details in
\link{read_grid}.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\value{
A \code{data.frame} object
}
\description{
Download annual estimates of access to employment, health and education
services by transport mode, as well as data on the spatial distribution of
population, schools and healthcare facilities at a fine spatial resolution
for all cities included in the study.
}
\examples{
# Read accessibility estimates of a single city
for <- read_access(city = 'Fortaleza', mode = 'walk', year = 2019, showProgress = FALSE)
for <- read_access(city = 'for', mode = 'walk', year = 2019, showProgress = FALSE)

# Read accessibility estimates for all cities
# all <- read_access(city = 'all', mode = 'public_transport', year = 2019)

}
\concept{accessibility data functions}
