% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_access.R
\name{read_access}
\alias{read_access}
\title{Download accessibility estimates}
\usage{
read_access(city, mode, year = 2019, showProgress = TRUE)
}
\arguments{
\item{city}{A city name or three-letter abbreviation}

\item{mode}{A transport mode.}

\item{year}{A date number in YYYY format.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\value{
A \code{data.frame} object
}
\description{
Download accessibility estimates
}
\examples{
\donttest{
# Read accessibility estimates of a single city
spo <- read_access(city = 'sao paulo', mode = 'walk', year = 2019)

# Read accessibility estimates of all cities in the project
spo <- read_access(city = 'all', mode = 'public_transport', year = 2019)
}
}
\concept{accessibility data functions}
