% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grid.R
\name{read_grid}
\alias{read_grid}
\title{Download spatial hexagonal grid}
\usage{
read_grid(city, showProgress = TRUE)
}
\arguments{
\item{city}{A city name or three-letter abbreviation}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\value{
An \verb{sd data.frame} object
}
\description{
Download spatial hexagonal grid
}
\examples{
\donttest{
# Read spatial grid of a single city
nat <- read_grid(city = 'natal')
nat <- read_grid(city = 'nat')

# Read spatial grid of all cities in the project
all <- read_grid(city = 'all')
}
}
\concept{spatial grid data functions}
