% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grid.R
\name{read_grid}
\alias{read_grid}
\title{Download spatial hexagonal grid H3}
\usage{
read_grid(city, showProgress = TRUE)
}
\arguments{
\item{city}{Character. A city name or three-letter abbreviation. If
\code{city="all"}, results for all cities are loaded.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\value{
An \verb{sf data.frame} object
}
\description{
Results of the AOP project are spatially aggregated on a hexagonal grid based
on the global H3 index at resolution 8, with a size of 357 meters (short
diagonal) and an area of 0.74 km2. More infor about H3 at
\url{https://h3geo.org/docs/core-library/restable/}.
}
\examples{
\donttest{
# Read spatial grid of a single city
nat <- read_grid(city = 'natal')
nat <- read_grid(city = 'nat')

# Read spatial grid of all cities in the project
all <- read_grid(city = 'all')
}
}
\concept{spatial data functions}
