% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_landuse.R
\name{read_landuse}
\alias{read_landuse}
\title{Download population and land use data}
\usage{
read_landuse(city = "bel", year = 2019, geometry = FALSE, showProgress = TRUE)
}
\arguments{
\item{city}{Character. A city name or three-letter abbreviation. If
\code{city="all"}, results for all cities are loaded.}

\item{year}{Numeric. A year number in YYYY format. Default set to 2019, the
only year currently available.}

\item{geometry}{Logical. If \code{FALSE} (the default), returns a regular data.table
of aop data. If \code{TRUE}, returns a an \verb{sf data.frame} with simple
feature geometry of spatial hexagonal grid H3. See details in
\link{read_grid}.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\value{
A \code{data.frame} object
}
\description{
Download data on the spatial distribution of population, schools and
healthcare facilities at a fine spatial resolution for cities included in the
study.
}
\examples{
\donttest{
# a single city
bho <- read_landuse(city = 'Belo Horizonte', year = 2019)
bho <- read_landuse(city = 'bho', year = 2019)

# all cities
all <- read_landuse(city = 'all', year = 2019)
}
}
\concept{land use data functions}
