% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_accent}
\alias{rm_accent}
\title{Remove accents from string}
\usage{
rm_accent(str, pattern = "all")
}
\arguments{
\item{str}{A string}

\item{pattern}{A pattern}
}
\description{
Remove accents from string
}
\seealso{
Other support functions: 
\code{\link{aop_spatial_join}()},
\code{\link{download_data}()},
\code{\link{load_data}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
