% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_city_input}
\alias{select_city_input}
\title{Select city input}
\usage{
select_city_input(temp_meta = temp_meta, city = NULL)
}
\arguments{
\item{temp_meta}{A dataframe with the file_url addresses of geobr datasets}

\item{city}{city input (passed from read_ function)}
}
\description{
Select city input
}
\seealso{
Other support functions: 
\code{\link{aop_spatial_join}()},
\code{\link{download_gpkg}()},
\code{\link{load_gpkg}()},
\code{\link{rm_accent}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
