% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_city_input}
\alias{select_city_input}
\title{Select city input}
\usage{
select_city_input(temp_meta = temp_meta, city = NULL)
}
\arguments{
\item{temp_meta}{A data.frame with the file_url addresses of aop datasets}

\item{city}{city input (passed from read_ function)}
}
\value{
A \code{data.frame} object with metadata subsetted by 'city'
}
\description{
Subsets the metadata table by 'city'.
}
\seealso{
Other support functions: 
\code{\link{aop_merge}()},
\code{\link{aop_spatial_join}()},
\code{\link{aopdata_dictionary}()},
\code{\link{check_connection}()},
\code{\link{download_data}()},
\code{\link{load_data}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
