% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_metadata}
\alias{select_metadata}
\title{Select metadata}
\usage{
select_metadata(t = NULL, c = NULL, y = NULL, m = NULL)
}
\arguments{
\item{t}{Type of data: 'access' or 'grid' (passed from \code{read_} function)}

\item{c}{City (passed from \code{read_} function)}

\item{y}{Year of the dataset (passed from \code{read_} function)}

\item{m}{Transport mode (passed from \code{read_} function)}
}
\description{
Select metadata
}
\examples{
\donttest{

library(geobr)

df <- download_metadata()

}

}
\seealso{
Other support functions: 
\code{\link{download_gpkg}()},
\code{\link{load_gpkg}()},
\code{\link{rm_accent}()},
\code{\link{select_city_input}()},
\code{\link{select_mode_input}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
