% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_mode_input}
\alias{select_mode_input}
\title{Select mode input}
\usage{
select_mode_input(temp_meta = temp_meta, mode = NULL)
}
\arguments{
\item{temp_meta}{A data.frame with the url addresses of aop datasets}

\item{mode}{Transport mode (passed by read_ function)}
}
\value{
A \code{data.frame} object with metadata subsetted by 'mode'
}
\description{
Subsets the metadata table by 'mode'.
}
\keyword{internal}
