% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_year_input}
\alias{select_year_input}
\title{Select year input}
\usage{
select_year_input(temp_meta = temp_meta, year = NULL)
}
\arguments{
\item{temp_meta}{A data.frame with the file_url addresses of aop datasets}

\item{year}{Year of the dataset (passed from read_ function)}
}
\value{
A \code{data.frame} object with metadata subsetted by 'year'
}
\description{
Subsets the metadata table by 'year'.
}
\keyword{internal}
