% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_year_input}
\alias{select_year_input}
\title{Select year input}
\usage{
select_year_input(temp_meta = temp_meta, year = NULL)
}
\arguments{
\item{temp_meta}{A dataframe with the file_url addresses of geobr datasets}

\item{year}{Year of the dataset (passed from read_ function)}
}
\description{
Select year input
}
\seealso{
Other support functions: 
\code{\link{aop_spatial_join}()},
\code{\link{download_gpkg}()},
\code{\link{load_gpkg}()},
\code{\link{rm_accent}()},
\code{\link{select_city_input}()},
\code{\link{select_metadata}()},
\code{\link{select_mode_input}()}
}
\concept{support functions}
